Added new exception types to ease error handling, fixed stylecop rules and stylecop violations, and updated tests.
Notable changes:
+ Added two new RequestException types with more to come:
AuthenticationFailureException and ResourceNotFoundException.
+ Moved Resource.EnsureStatusCodeAsnyc to Response.EnsureStatusCodeAsync
so that it can be used whenever a response is in hand.
+ Changed IndexAttribute property types to ensure that if a value is
set, it is transmitted. This is to support attribute updates where it
may be desirable to set the value back to its default by specifying the
default value explicitly. If a value is set (i.e., it is non-null) it is
transmitted.