Bug fixes
* DVPL-4448: Service.GetIndexesAsync().Result call causes the stack
overflow
Service.Namespace was calling itself recursively. I switched to backing
fields for readonly properties in Service.cs and missed a recursive
reference to Service.Namespace.

* DVPL-4438: Index.UpdateAsync() call throws exception
I was checking for HttpStatusCode.Created, not HttpStatusCode.OK. We now
have a test that verifies we can create an index and update it. Next up:
get and then remove the updated index.