DVPL-4443: SavedSearch.DispatchAsync() result can't be parsed
SavedSearch.DispatchAsync is now using the correct resource name and
checking for the correct HttpStatusCode. I posted a note on Splunk.com
that the documenation for the [POST
saved/searches/{name}/dispatch|http://goo.gl/AfzBJO] endpoint is
incorrect. It claims that the endpoint returns HttpStatusCode.OK when in
fact it returns HttpStatus.Created. Note the documentation improvements.