Discovered and fixed three bugs
* StoragePassword: ResourceName.Title is now computed properly based on
username and realm: "<realm>:<username>:".
* Example: ":admin:", if realm is null or empty.
* Example: "splunk.com:admin:", if realm is "splunk.com".
* Example: "splunk\:com:david\:noble", if realm is "splunk:com" and
username is "david:noble".

* ResourceName.CompareTo returns zero when resource name are equal
instead of throwing an exception. NOTE: System.Linq.Enumerable.First is
incorrectly documented in source code comments, indicating that a value
of default(TSource) is returned when no match to the predicate is found.

* Resource instances are now compared using ResourceName only. This is
because 1) DNS names cannot be reliably compared and 2) The namespace
used to locate a resource has nothing to do with identity, but rather
visibility.