Resolved DVPL-5630, DVPL-6736, and DVPL-6737
* DVPL-5630, C# SDK - ExportSearchResultsAsync error when no results
* DVPL-6736, C# SDK | Service.ExportSearchResultsAsync and Service.ExportSearchPreviewsAsync do not recognize
  errors that occur during export
* DVPL-6737, C# SDK | Splunk 6.3 | Address breaks in ServerInfo class

The fix to DVPL-6737 changed ServerInfo.Build from an int to a long. This was required by a change to the length
and meaning of build numbers in Splunk 6.3. They used to be decimal serial numbers. They are now 12-byte hexadecimal
Git commit numbers. As such build numbers should only be compared for equality. They are no longer serial numbers.