Handle cases where we get an empty key for an s:dict entry
When EAI sends us s:dict entries (with nested s:key nodes), we parse
them as a dictionary in C#. However, it turns out the API can send us
a case where the value for the s:key node is actually the empty string,
which we do not expect and actually have a code contract to protect
against.

In that case, we will change it from the empty string to the string
called "empty" to be explicit about it. Since there can only be one
empty key per level (otherwise it would be a serious API error), this
is a reasonable workaround.